Rename: 게시물 디렉토리 이름을 board로 변경
