+++ src/main/java/com/greeny/ecomate/board/dto/CreateBoardRequestDto.java
@@ -1,4 +1,4 @@
-package com.greeny.ecomate.posting.dto;
+package com.greeny.ecomate.board.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
