+++ src/main/java/com/greeny/ecomate/comment/repository/CommentRepository.java
@@ -1,7 +1,7 @@
 package com.greeny.ecomate.comment.repository;
 
 import com.greeny.ecomate.comment.entity.Comment;
-import com.greeny.ecomate.posting.entity.Board;
+import com.greeny.ecomate.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
