+++ src/main/java/com/greeny/ecomate/comment/service/CommentService.java
@@ -6,8 +6,8 @@ import com.greeny.ecomate.comment.dto.CreateCommentRequestDto;
 import com.greeny.ecomate.comment.entity.Comment;
 import com.greeny.ecomate.comment.repository.CommentRepository;
 import com.greeny.ecomate.exception.NotFoundException;
-import com.greeny.ecomate.posting.entity.Board;
-import com.greeny.ecomate.posting.repository.BoardRepository;
+import com.greeny.ecomate.board.entity.Board;
+import com.greeny.ecomate.board.repository.BoardRepository;
 import com.greeny.ecomate.member.entity.Member;
 import com.greeny.ecomate.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
