+++ src/main/java/com/greeny/ecomate/like/entity/Like.java
@@ -1,9 +1,8 @@
 package com.greeny.ecomate.like.entity;
 
 import com.greeny.ecomate.base.BaseEntity;
-import com.greeny.ecomate.posting.entity.Board;
+import com.greeny.ecomate.board.entity.Board;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
