+++ src/main/java/com/greeny/ecomate/like/repository/LikeRepository.java
@@ -1,7 +1,7 @@
 package com.greeny.ecomate.like.repository;
 
 import com.greeny.ecomate.like.entity.Like;
-import com.greeny.ecomate.posting.entity.Board;
+import com.greeny.ecomate.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
